+++ hypervisor/arch/x86/guest/vm.c
@@ -671,8 +671,8 @@ int32_t create_vm(uint16_t vm_id, uint64_t pcpu_bitmap, struct acrn_vm_config *v
 	}
 
 	if (status == 0) {
-		uint32_t i;
-		for (i = 0; i < vm_config->pt_intx_num; i++) {
+		uint16_t i;
+		for (i = 0U; i < vm_config->pt_intx_num; i++) {
 			status = ptirq_add_intx_remapping(vm, vm_config->pt_intx[i].virt_gsi,
 								vm_config->pt_intx[i].phys_gsi, false);
 			if (status != 0) {
